requirements.txt
pywizlight
pyyaml
colour
asyncio_dgram
pynput
python-kasa
